test module breakdown + indentation split
